+++ src/metrics/loc/ruby.rs
@@ -73,7 +73,7 @@ impl Loc for RubyCode {
             | R::Undef
             | R::Alias
             | R::EmptyStatement => {
-                stats.lloc.logical_lines += 1;
+                stats.lloc.count_logical_line();
             }
             _ => {
                 check_comment_ends_on_code_line(stats, start);